Re-added Birch example pokemon randomization
